load all data to local

+++ app/src/main/java/com/asiasquare/byteg/shoppingdemo/cart/CartFragment.kt
@@ -1,17 +1,22 @@
 package com.asiasquare.byteg.shoppingdemo.cart
 
 
+import android.content.DialogInterface
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.lifecycleScope
+import androidx.recyclerview.widget.ItemTouchHelper
+import androidx.recyclerview.widget.RecyclerView
 import com.asiasquare.byteg.shoppingdemo.database.items.ShoppingBasketItem
 import com.asiasquare.byteg.shoppingdemo.databinding.FragmentCartBinding
-import com.asiasquare.byteg.shoppingdemo.favorite.FavoriteFragmentAdapter
+import com.google.android.material.snackbar.Snackbar
+import kotlinx.coroutines.flow.collect
 
 class CartFragment : Fragment(), CartFragmentAdapter.OnClickListener {
 
@@ -49,9 +54,47 @@ class CartFragment : Fragment(), CartFragmentAdapter.OnClickListener {
         /** Update data to adapter **/
         viewModel.cartList.observe(viewLifecycleOwner, Observer {
             it?.let {
+                if (it.isEmpty()) {
+                    binding.emptyView.visibility = View.VISIBLE
+                    binding.buttonDatHang.visibility = View.GONE
+                } else
+                    binding.emptyView.visibility = View.GONE
                 adapter.submitList(it)
             }
         })
+
+        /** Swipe delete functionality **/
+        ItemTouchHelper(object : ItemTouchHelper.SimpleCallback(
+            0,
+            ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT
+        ) {
+            override fun onMove(
+                recyclerView: RecyclerView,
+                viewHolder: RecyclerView.ViewHolder,
+                target: RecyclerView.ViewHolder
+            ): Boolean {
+                return false
+            }
+
+            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+                val task = adapter.currentList[viewHolder.absoluteAdapterPosition]
+                viewModel.onTaskSwiped(task)
+            }
+        }).attachToRecyclerView(binding.recyclerViewGioHang)
+
+        /** Undo delete functionality **/
+        viewLifecycleOwner.lifecycleScope.launchWhenStarted {
+            viewModel.tasksEvent.collect { event ->
+                when (event) {
+                    is CartFragmentViewModel.TasksEvent.ShowUndoDeleteTaskMessage -> {
+                        Snackbar.make(requireView(), "Đã xóa sản phẩm", Snackbar.LENGTH_LONG)
+                            .setAction("Hoàn tác") {
+                                viewModel.onUndoDeleteClick(event.task)
+                            }.show()
+                    }
+                }
+            }
+        }
         return binding.root
     }
 
@@ -61,8 +104,38 @@ class CartFragment : Fragment(), CartFragmentAdapter.OnClickListener {
         _binding = null
     }
 
+
     override fun onDeleteClick(cart: ShoppingBasketItem) {
-        viewModel.onDeleteCartClicking(cart)
+        val dialogBuilder = AlertDialog.Builder(requireActivity())
+
+        // set message of alert dialog
+        dialogBuilder.setMessage("Bạn có chắc chắn muốn xóa sản phẩm này không?")
+            // if the dialog is cancelable
+            .setCancelable(false)
+            // positive button text and action
+            .setPositiveButton("Xác nhận", DialogInterface.OnClickListener { dialog, id ->
+                viewModel.onDeleteCartClicking(cart)
+            })
+            // negative button text and action
+            .setNegativeButton("Quay lại", DialogInterface.OnClickListener { dialog, id ->
+                dialog.cancel()
+            })
+
+        // create dialog box
+        val alert = dialogBuilder.create()
+        // set title for alert dialog box
+        alert.setTitle("Xác nhận xóa")
+        // show alert dialog
+        alert.show()
+
+    }
+
+    override fun onAddClick(cart: ShoppingBasketItem) {
+        viewModel.onAddBtnClicking(cart)
+    }
+
+    override fun onMinusClick(cart: ShoppingBasketItem) {
+        viewModel.onMinusBtnClicking(cart)
     }
